UIU-2143 include missing limit clause in request-count query (#1714)
The limit clause in the request count query was missing. We did a nice
job batching the query to handle a long list of loan-ids, but,
astonishingly, completely forgot that if we had, say, 15 loan-ids, we
might get 15+ requests (or even if we had only one loan-id that it might
have 15 open requests). Whoops.

Refs UIU-2143