UIU-2163 avoid querying for an empty list of loan policies
A list of unique loan policy IDs is extracted from the list of loans,
but that list can still be empty. (This does not really make sense to
me, but that's the data we have: sometimes, loans don't have loan
policies.) When the list is empty, we don't need to do anything.
Previously, we would send an invalid query, `id==()`.

Refs UIU-2163