UIU-2158 use correct prop-types
Man oh man what a PITA `react-intl` `v5` is when it comes to
render-props. Sheesh. Fix that shiz mainly by using `injectIntl` and
`formatMessage` to get strings where they are necessary; it's so much
cleaner and easier to follow.

And sync other prop-types that have drifted out of sync and are now
causing console pollution.

Refs UIU-2158