Fixed accordion open/close in user form (#1771)
The Expand All button wasn't working in the User form because
the proxy accordion had the wrong ID/state pairing. I went
ahead and just ported it over to the new `AccordionStatus`
component so that the individual accordion state didn't need
to be kept track of and risk stuff like this again.