UIU-1971: Delete user with transactions check (#1768)
Allow patrons to be removed if they have no open transactions
(loans, requests, fines/fees, blocks, unexpired proxy) as determined
by a request to `bl-users/by-id/:{id}/open-transactions`. 

The remove-user option is provided in the actions menu of 
the patron-details pane and is visible only to properly 
permissioned users. Show a confirmation modal before
proceeding with removal; show a "close these transactions" 
modal if the patron cannot be removed. 

Refs UIU-1971, MODUSERBL-115