UIU-2435 don't choke when editing minimal user
It's possible to create a minimal user object at the `/users` endpoint,
e.g. a POST request with a username field and nothing else, but editing
such a record in the UI fails because fields the UI requires (though the
backend does not) are missing.

Provide default value (`"false"`) for the `active` field when it is
missing.

Refs UIU-2435