UIU-2449 do not label loan-less fees as "Anonymized" (#1922)
Only fees/fines that are associated with now-anonymized loans should
display "Anonymized" under the "Loan details" label. The previous check
for this condition was naive, interrogating only the presence of a
`loanId` which happens to be empty after anonymization ... but also
empty for fees/fines that never were associated with loans in the first
place.

Refs UIU-2449