UIU-2409 confirm editing of 'visible: false' permissions
The permissions dialog only shows permissions with `visible: true`
attributes, but sometimes permissions with `visible: false` are assigned
to users. Editing such a user or permission set will cause the `visible:
false` values to be removed (see UIU-2410).

Show a prompt in this situation explaining the situation so the user can
choose whether to continue with the edit action, losing the hidden
permissions, or cancel, leaving things as they are.

Refs UIU-2409