UIU-2409 refactor for better test coverage
Refactor `...List` and `...ListItem` into their own components so they
can be adequately covered with tests. I think the lesson here is
basically that any private function returning a component is best
refactored into a stand-alone component. If it's not exported, then it's
still effectively private, but because it's separated from its container
it becomes dramatically easier to test.

Refs UIU-2409