remove broken OverrideModal test
This test is broken, probably because the `id` selector is brittle
and accesses a dynamically-created element, so probably something like
```
  const input = screen.getByLabelText('ui-users.blocks.modal.comment');
  userEvent.type(input, 'testInfoModified');
```
would be better. Additionally, it's unclear where the expected value
"TestInfot" comes from, nor why it would be valuable to search the document
for text that was just typed in. :shrug: