UIU-1900 unpin moment and move it to peer
Long ago and far away, we pinned `moment` to `~2.24.0` because of
problems in `2.25.0` (STRIPES-678) but that's since been resolved, and
there are now security warnings about this old version.

The real substantive change here is that `settings` now correctly
imports from `moment-timezone` instead of straight `moment`. Score
another one for the "Wait, how did this ever work?!?" team.

Refs UIU-1900