UIU-2631 accurately count loans > 1000 (#2191)
The open-loan count on the "Loans" accordion could be inaccurate when
there were more than 1000 loans because it used a `limit=1` query
parameter (which still allows for an [estimated `totalRecords`
value](https://github.com/folio-org/raml-module-builder#estimated-totalrecords)
instead of `limit=0` which forces an accurate count.

Display the badge value on the closed accordion with `<FormattedNumber>`
so it will be correctly formatted when the value is > 999.

Refs UIU-2631