UIU-2750: Upgrade yarn.lock fixing decode-uri-component DoS CVE-2022-38900
Upgrading yarn.lock upgrades decode-uri-component from 0.2.0 to 0.2.2 fixing Denial of Service (DoS):

https://www.cve.org/CVERecord?id=CVE-2022-38900

For production not this yarn.lock file but the platform yarn.lock file is used. However, the builds from this repository should contain the fixed decode-uri-component version so that any side-effects pop up early and we can be sure that ui-users is ready to use the fixed decode-uri-component version.