UIU-2803 show an error if proxy record is corrupt
Sometimes, it is possible to create a proxy record that has an invalid
`proxyUserId` attribute, resulting in `ProxyItem` receiving a `record`
prop with a null `user` attribute. It isn't totally clear how this
happens, but it clearly does happen sometimes, and so the UI needs to
handle the situation. e.g. it is possible to create a corrupt record by
directly sending a PUT request to `/proxiesfor/${id}` with invalid data,
even if it's not possible to create such a request in the UI.

Refs UIU-2803