UIU-2901 handle sparse borrower data in overdue loans report
If a user record attached to a loan in the overdue loans report has been
removed, the `userId` attribute will be present but the `borrower`
object will not be. The parser needs to be robust to this situation.

Curiously, such an NPE causes parsing to silently halt: there is no
warning in the console, no stacktrace, nothing at all. `map` just never
returns. In local tests, I get an NPE in similar situations, so I don't
understand how the error gets swallowed here, but it does.

Refs UIU-2901