include @aashutoshrathi/word-wrap as a dev-dep (#2657)
Include `@aashutoshrathi/word-wrap` as a dev-dep in addition to
including it in `resolutions` where it is used as an alias to replace
the (unmaintained) `word-wrap`. Just like peer-deps, yarn does not satisfy
resolution-deps, so if we want that reference in `resolutions` to be
resolvable, we need to _actually_ depend on it elsewhere. `word-wrap` is a
transitive dep via eslint > optionator, so I added
`@aashutoshrathi/word-wrap` as a dev-dep.

Additionally, unrelated, it is important to leave `--frozen-lockfile` in
the `Run yarn install` instructions. Without it, there's not much point
in committing `yarn.lock` because you would be free to update
`package.json` by changing versions or even adding new dependencies and
CI wouldn't warn you that `yarn.lock` and `package.json` were out of
sync. This breaks the promise of reprodicble builds that committing
`yarn.lock` provides.