Remove two vacuous tests
The tests are that when the patron is virtual, the Add Patron Info and
Add Staff Info buttons should be disabled. But in fact they do not get
rendered at all, presumably due to the new `<InInterface>` wrapping
being incorrectly mocked.

I spent some time trying to figure our what was going wrong with the
mocking before realising that the tests are fundamentally valueless
and this was not a good use of my time.