[UIU-3406] Loans accordion shows counts use-at-location loans (#2928)
When displaying the full information about a user in the right page, the Loans accordion, when expanded, now shows not only the counts of open and closed loans, but within the former also shows information about use-at-location loans, if there are any. The information breaks down the count by how many of them are "Held" and "In use".